requirements.txt
pandas
numpy
torch==2.3.0
transformers
hydra-core
pyvi
torchtext